Consolidate all applications.
This is a keycloak POC to see how we can integrate keycloak with an
angular frontend, a spring boot based backend, an event provider to send
registration events to a kafka topic so we can do extra work.

All the components have been moved to this repo to ensure that any user
can set it up quickly and effortlessly.